src/main/python/requirements.txt
hazelcast-python-client
sqlalchemy
testcontainers
